+++ autotests/testscreenconfig.cpp
@@ -109,7 +109,7 @@ void testScreenConfig::singleOutput()
 
     const ModePtr mode = output->auto_mode();
     QCOMPARE(mode->size(), QSize(1280, 800));
-    QCOMPARE(mode->refreshRate(), (float)59.9);
+    QCOMPARE(mode->refreshRate(), 59.9);
 }
 
 void testScreenConfig::singleOutputWithoutPreferred()
@@ -160,7 +160,7 @@ void testScreenConfig::multiOutput()
     const ModePtr mode = output->auto_mode();
     QVERIFY(!mode.isNull());
     QCOMPARE(mode->size(), QSize(1920, 1080));
-    QCOMPARE(mode->refreshRate(), (float)60.0);
+    QCOMPARE(mode->refreshRate(), 60.0);
 }
 
 void testScreenConfig::configCanBeApplied()
